+++ lib/src/timeconsuming_page.dart
@@ -1,43 +1,60 @@
 part of timeconsuming_page_builder;
 
+/// Signature for adding ability to retry when error occurred.
+///
+/// Used by [ErrorWidgetBuilder]
 typedef RetryCaller = void Function();
 
+/// Signature for building waiting widget.
+/// See also:
+///  * [BuiltInWaitingWidget], which is built-in waiting widget.
 typedef WaitingWidgetBuilder = Widget Function(BuildContext context);
 
+/// Signature for building error widget.
+/// See also:
+///  * [BuiltInErrorWidget], which is built-in error widget.
 typedef ErrorWidgetBuilder = Widget Function(
     BuildContext context, RetryCaller caller);
 
+/// Signature for building data widget.
+/// See also:
+///  * [BuiltInEmptyWidget], which is built-in empty widget.
 typedef DataWidgetBuilder<T> = Widget Function(BuildContext context, T data);
 
+/// Signature for building [Future].
 typedef FutureBuilder<T> = Future<T> Function();
 
-class TimeConsumingPage<T> extends StatefulWidget {
-  static AsyncWidgetBuilder<T> _asyncWidgetBuilder<T>({
-    @required WaitingWidgetBuilder waitingWidgetBuilder,
-    @required ErrorWidgetBuilder errorWidgetBuilder,
-    @required DataWidgetBuilder<T> dataWidgetBuilder,
-  }) {
-    return (BuildContext context, AsyncSnapshot<T> snapshot) {
-      debugPrint("[TimeConsumingPage] $snapshot");
-      Widget child;
-      if (snapshot.hasData) {
-        child = dataWidgetBuilder(context, snapshot.data);
-      } else if (snapshot.hasError) {
-        child = errorWidgetBuilder(context, () {
-          final state =
-              (context as StatefulElement).state as _TimeConsumingPageState;
-          state.resubscribe();
-        });
-      } else {
-        child = waitingWidgetBuilder(context);
-      }
-      return child;
-    };
-  }
-
-  TimeConsumingPage({
+/// Widgets that switches widget state based on different situation
+/// on a time-consuming page with a [Future].
+///
+/// ```dart
+/// TimeConsumingPageBuilder<int>(
+///   futureBuilder: ()=> Future.delayed(Duration(microseconds: 10), () => 8),
+///   waitingWidgetBuilder: (BuildContext context) {
+///     // show waiting to user.
+///     return ...; //
+///   },
+///   errorWidgetBuilder: (BuildContext context, RetryCaller caller) {
+///     // show error to the user.
+///     // caller() will retry again.
+///     return ...;
+///   },
+///   dataWidgetBuilder: (BuildContext context, int data) {
+///     // show data to the user.
+///     // parameter[data] will be set 8 after 10 microseconds.
+///     return ...;
+///   },
+/// )
+/// ```
+class TimeConsumingPageBuilder<T> extends StatefulWidget {
+  /// Creates a time-consuming page builder.
+  ///
+  /// [futureBuilder] which used to build future.
+  /// [waitingWidgetBuilder] which used to build waiting widget.
+  /// [errorWidgetBuilder] which used to build error widget.
+  /// [dataWidgetBuilder] which used to build data widget.
+  TimeConsumingPageBuilder({
     Key key,
-    this.initialData,
     @required this.futureBuilder,
     @required WaitingWidgetBuilder waitingWidgetBuilder,
     @required ErrorWidgetBuilder errorWidgetBuilder,
@@ -52,31 +69,53 @@ class TimeConsumingPage<T> extends StatefulWidget {
             dataWidgetBuilder: dataWidgetBuilder),
         super(key: key);
 
+  static AsyncWidgetBuilder<T> _asyncWidgetBuilder<T>({
+    @required WaitingWidgetBuilder waitingWidgetBuilder,
+    @required ErrorWidgetBuilder errorWidgetBuilder,
+    @required DataWidgetBuilder<T> dataWidgetBuilder,
+  }) {
+    return (BuildContext context, AsyncSnapshot<T> snapshot) {
+      debugPrint("[TimeConsumingPage] $snapshot");
+      Widget child;
+      if (snapshot.hasData) {
+        child = dataWidgetBuilder(context, snapshot.data);
+      } else if (snapshot.hasError) {
+        child = errorWidgetBuilder(context, () {
+          final state = (context as StatefulElement).state
+              as _TimeConsumingPageBuilderState;
+          state.resubscribe();
+        });
+      } else {
+        child = waitingWidgetBuilder(context);
+      }
+      return child;
+    };
+  }
+
   final FutureBuilder<T> futureBuilder;
 
   final AsyncWidgetBuilder<T> builder;
 
-  final T initialData;
-
   @override
-  State<TimeConsumingPage<T>> createState() => _TimeConsumingPageState<T>();
+  State<TimeConsumingPageBuilder<T>> createState() =>
+      _TimeConsumingPageBuilderState<T>();
 }
 
-/// State for [TimeConsumingPage].
-class _TimeConsumingPageState<T> extends State<TimeConsumingPage<T>> {
+/// State for [TimeConsumingPageBuilder].
+class _TimeConsumingPageBuilderState<T>
+    extends State<TimeConsumingPageBuilder<T>> {
   Object _activeCallbackIdentity;
   AsyncSnapshot<T> _snapshot;
 
   @override
   void initState() {
     super.initState();
-    _snapshot =
-        AsyncSnapshot<T>.withData(ConnectionState.none, widget.initialData);
+    _snapshot = AsyncSnapshot<T>.withData(ConnectionState.none, null);
     _subscribe();
   }
 
   @override
-  void didUpdateWidget(TimeConsumingPage<T> oldWidget) {
+  void didUpdateWidget(TimeConsumingPageBuilder<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.futureBuilder != widget.futureBuilder) {
       if (_activeCallbackIdentity != null) {
@@ -99,8 +138,7 @@ class _TimeConsumingPageState<T> extends State<TimeConsumingPage<T>> {
   void resubscribe() {
     if (_activeCallbackIdentity != null) {
       _unsubscribe();
-      _snapshot =
-          AsyncSnapshot<T>.withData(ConnectionState.none, widget.initialData);
+      _snapshot = AsyncSnapshot<T>.withData(ConnectionState.none, null);
     }
     _subscribe(notifyWaitingState: true);
   }