+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+Version 1.1.0 *(2020-09-01)*
+----------------------------
+
+### Added
+- Add comments.
+### Changed
+- Rename `TimeConsumingPage` to `TimeConsumingPageBuilder`.
+- Remove named constructor parameter: `initialData`.
+
 Version 1.0.0 *(2020-09-01)*
 ----------------------------